+++ packages/react/src/Deferred.ts
@@ -1,4 +1,4 @@
-import { ReactElement, ReactNode, useEffect, useMemo, useState } from 'react'
+import { ReactNode, useEffect, useMemo, useState } from 'react'
 import { router } from '.'
 import usePage from './usePage'
 
@@ -15,7 +15,7 @@ const isSameUrlWithoutHash = (url1: URL | Location, url2: URL | Location): boole
 
 interface DeferredProps {
   children: ReactNode | (() => ReactNode)
-  fallback: ReactElement | number | string
+  fallback: ReactNode | (() => ReactNode)
   data: string | string[]
 }
 
@@ -47,7 +47,9 @@ const Deferred = ({ children, data, fallback }: DeferredProps) => {
     setLoaded(keys.every((key) => pageProps[key] !== undefined))
   }, [pageProps, keys])
 
-  if (!loaded) return fallback
+  if (!loaded) {
+    return fallback
+  }
 
   return typeof children === 'function' ? children() : children
 }