+++ tests/deferred-props.spec.ts
@@ -42,6 +42,8 @@ test('we are not caching deferred props after reload', async ({ page }) => {
 
   await expect(page.getByText('Loading foo...')).toBeVisible()
   await expect(page.getByText('Loading bar...')).toBeVisible()
+  await expect(page.getByText('foo value')).not.toBeVisible()
+  await expect(page.getByText('bar value')).not.toBeVisible()
 
   await page.waitForResponse(page.url())
 
@@ -54,6 +56,8 @@ test('we are not caching deferred props after reload', async ({ page }) => {
 
   await expect(page.getByText('Loading foo...')).toBeVisible()
   await expect(page.getByText('Loading bar...')).toBeVisible()
+  await expect(page.getByText('foo value')).not.toBeVisible()
+  await expect(page.getByText('bar value')).not.toBeVisible()
 
   await page.waitForResponse(page.url())